Comments were Added [v0.1]

+++ OnlineTestManagement/src/main/java/com/capgemini/onlinetestmanagement/dao/AssignExamToUserDaoI.java
@@ -18,6 +18,6 @@ public interface AssignExamToUserDaoI extends JpaRepository<AssignExamToUser,Int
 	
 	
 	@Query("from AssignExamToUser examuser inner join fetch examuser.exam e inner join fetch examuser.user u where u.userId=:userId")
-	 List<AssignExamToUser> getListOfExamsAssignToUser(int userId);
+	 List<AssignExamToUser> getListOfExamsAssignToUser(long userId);
 	
 }
